+++ test/unit/ask_for_help_bot/mock.py
@@ -41,3 +41,4 @@ class MockAskForHelpHandler(AskForHelpHandler):
         self.task_type_id = "task_type_id"
         self.community_id = "community_id"
         self.max_users = 5
+        self.survey_url = "survey_url"