+++ test/unit/ask_for_help_bot/test_handler.py
@@ -648,3 +648,20 @@ class TestAskForHelpHandler(TestCase):
         self.assertIsInstance(response, OutgoingEvent)
         self.assertEqual(1, len(response.messages))
         self.assertIsInstance(response.messages[0], TextualResponse)
+
+    def test_action_start(self):
+        handler = MockAskForHelpHandler()
+        translator_instance = TranslatorInstance("wenet-ask-for-help", None, handler._alert_module)
+        translator_instance.translate = Mock(return_value="")
+        handler._translator.get_translation_instance = Mock(return_value=translator_instance)
+        handler._get_user_locale_from_incoming_event = Mock(return_value="en")
+
+        response = handler.action_start(IncomingTelegramEvent("", TelegramDetails(1, 1, ""), IncomingCommand("message_id", int(datetime.now().timestamp()), "user_id", "chat_id", handler.INTENT_START, ""), ConversationContext()), handler.INTENT_START)
+        self.assertIsInstance(response, OutgoingEvent)
+        self.assertEqual(5, len(response.messages))
+        self.assertIsInstance(response.messages[0], TextualResponse)
+        self.assertIsInstance(response.messages[1], TextualResponse)
+        self.assertIsInstance(response.messages[2], TextualResponse)
+        self.assertIsInstance(response.messages[3], TextualResponse)
+        self.assertIsInstance(response.messages[4], TelegramRapidAnswerResponse)
+        self.assertEqual(1, len(response.messages[4].options))